requirements.txt
PyQt5~=5.15.1
pandas~=1.1.2
matplotlib~=3.3.2
python-dateutil~=2.8.1
setuptools~=47.1.0